+++ tests/core/test_library.py
@@ -5,7 +5,6 @@ import unittest
 import mock
 
 from mopidy import backend, core
-from mopidy.internal import deprecation
 from mopidy.models import Image, Ref, SearchResult, Track
 
 
@@ -151,10 +150,6 @@ class CoreLibraryTest(BaseCoreLibraryTest):
     def test_lookup_returns_empty_dict_for_no_uris(self):
         self.assertEqual({}, self.core.library.lookup(uris=[]))
 
-    def test_lookup_fails_with_uri_and_uris_set(self):
-        with self.assertRaises(ValueError):
-            self.core.library.lookup('dummy1:a', ['dummy2:a'])
-
     def test_lookup_can_handle_uris(self):
         track1 = Track(uri='dummy1:a', name='abc')
         track2 = Track(uri='dummy2:a', name='def')
@@ -278,34 +273,6 @@ class CoreLibraryTest(BaseCoreLibraryTest):
         self.core.library.search({'any': 'foobar'})
         self.library1.search.assert_called_once_with(
             query={'any': ['foobar']}, uris=None, exact=False)
-
-
-class DeprecatedLookupCoreLibraryTest(BaseCoreLibraryTest):
-
-    def run(self, result=None):
-        with deprecation.ignore('core.library.lookup:uri_arg'):
-            return super(DeprecatedLookupCoreLibraryTest, self).run(result)
-
-    def test_lookup_selects_dummy1_backend(self):
-        self.library1.lookup.return_value.get.return_value = []
-        self.core.library.lookup('dummy1:a')
-
-        self.library1.lookup.assert_called_once_with('dummy1:a')
-        self.assertFalse(self.library2.lookup.called)
-
-    def test_lookup_selects_dummy2_backend(self):
-        self.library2.lookup.return_value.get.return_value = []
-        self.core.library.lookup('dummy2:a')
-
-        self.assertFalse(self.library1.lookup.called)
-        self.library2.lookup.assert_called_once_with('dummy2:a')
-
-    def test_lookup_uri_returns_empty_list_for_dummy3_track(self):
-        result = self.core.library.lookup('dummy3:a')
-
-        self.assertEqual(result, [])
-        self.assertFalse(self.library1.lookup.called)
-        self.assertFalse(self.library2.lookup.called)
 
 
 class LegacyFindExactToSearchLibraryTest(unittest.TestCase):
@@ -468,24 +435,6 @@ class LookupByUrisBadBackendTest(MockBackendCoreLibraryBase):
         uri = 'dummy:/1'
         self.library.lookup.return_value.get.return_value = [123]
         self.assertEqual({uri: []}, self.core.library.lookup(uris=[uri]))
-        logger.error.assert_called_with(mock.ANY, 'DummyBackend', mock.ANY)
-
-    def test_backend_returns_none_with_uri(self, logger):
-        uri = 'dummy:/1'
-        self.library.lookup.return_value.get.return_value = None
-        self.assertEqual([], self.core.library.lookup(uri))
-        self.assertFalse(logger.error.called)
-
-    def test_backend_returns_wrong_type_with_uri(self, logger):
-        uri = 'dummy:/1'
-        self.library.lookup.return_value.get.return_value = 'abc'
-        self.assertEqual([], self.core.library.lookup(uri))
-        logger.error.assert_called_with(mock.ANY, 'DummyBackend', mock.ANY)
-
-    def test_backend_returns_iterable_wrong_types_with_uri(self, logger):
-        uri = 'dummy:/1'
-        self.library.lookup.return_value.get.return_value = [123]
-        self.assertEqual([], self.core.library.lookup(uri))
         logger.error.assert_called_with(mock.ANY, 'DummyBackend', mock.ANY)